+++ bridge/windows_process_attrs_windows.go
@@ -1,0 +1,11 @@
+//go:build windows
+
+// Windows process attributes keep helper updater scripts out of the foreground.
+
+package main
+
+import "syscall"
+
+func windowsHiddenProcessAttrs() *syscall.SysProcAttr {
+	return &syscall.SysProcAttr{HideWindow: true}
+}